+++ tests/test_util.py
@@ -1,6 +1,19 @@
 import pytest
 
-from antidote._utils import get_arguments_specification
+from antidote._utils import Argument, ArgumentSpecification, SlotReprMixin, \
+    get_arguments_specification
+
+
+class DummySlot(SlotReprMixin):
+    __slots__ = ('test', 'value')
+
+    def __init__(self, test, value):
+        self.test = test
+        self.value = value
+
+
+def test_slot_repr_mixin():
+    assert repr(DummySlot(1, 'test')) == "DummySlot(test=1, value='test')"
 
 
 def f(a, b, c=1):
@@ -19,7 +32,7 @@ def k():
     pass
 
 
-class Dummy(object):
+class Dummy:
     def f(self, a, b=1, *args, **kwargs):
         pass
 
@@ -38,22 +51,123 @@ d = Dummy()
 @pytest.mark.parametrize(
     'func,expected',
     [
-        (f, ([('a', False), ('b', False), ('c', True)], False, False)),
-        (g, ([('a', False)], True, False)),
-        (h, ([('b', True)], False, True)),
-        (k, ([], False, False)),
-        (Dummy.f, ([('self', False), ('a', False), ('b', True)], True, True)),
-        (Dummy.g, ([('a', False)], False, False)),
-        (Dummy.h, ([], False, False)),
-        (d.f, ([('a', False), ('b', True)], True, True)),
-        (d.g, ([('a', False)], False, False)),
-        (d.h, ([], False, False)),
-    ],
-    ids=[
-        'f', 'g', 'h', 'k',
-        'cls.f', 'cls.g', 'cls.h',
-        'instance.f', 'instance.g', 'instance.h'
+        pytest.param(
+            f,
+            ArgumentSpecification(
+                arguments=tuple([
+                    Argument('a', False),
+                    Argument('b', False),
+                    Argument('c', True),
+                ]),
+                has_var_positional=False,
+                has_var_keyword=False,
+            ),
+            id='f'
+        ),
+        pytest.param(
+            g,
+            ArgumentSpecification(
+                arguments=tuple([
+                    Argument('a', False),
+                ]),
+                has_var_positional=True,
+                has_var_keyword=False,
+            ),
+            id='g'
+        ),
+        pytest.param(
+            h,
+            ArgumentSpecification(
+                arguments=tuple([
+                    Argument('b', True),
+                ]),
+                has_var_positional=False,
+                has_var_keyword=True,
+            ),
+            id='h'
+        ),
+        pytest.param(
+            k,
+            ArgumentSpecification(
+                arguments=tuple([]),
+                has_var_positional=False,
+                has_var_keyword=False,
+            ),
+            id='k'
+        ),
+        pytest.param(
+            Dummy.f,
+            ArgumentSpecification(
+                arguments=tuple([
+                    Argument('self', False),
+                    Argument('a', False),
+                    Argument('b', True),
+                ]),
+                has_var_positional=True,
+                has_var_keyword=True,
+            ),
+            id='cls.f'
+        ),
+        pytest.param(
+            Dummy.g,
+            ArgumentSpecification(
+                arguments=tuple([
+                    Argument('a', False),
+                ]),
+                has_var_positional=False,
+                has_var_keyword=False,
+            ),
+            id='cls.g'
+        ),
+        pytest.param(
+            Dummy.h,
+            ArgumentSpecification(
+                arguments=tuple([]),
+                has_var_positional=False,
+                has_var_keyword=False,
+            ),
+            id='cls.h'
+        ),
+        pytest.param(
+            d.f,
+            ArgumentSpecification(
+                arguments=tuple([
+                    Argument('a', False),
+                    Argument('b', True),
+                ]),
+                has_var_positional=True,
+                has_var_keyword=True,
+            ),
+            id='instance.f'
+        ),
+        pytest.param(
+            d.g,
+            ArgumentSpecification(
+                arguments=tuple([
+                    Argument('a', False),
+                ]),
+                has_var_positional=False,
+                has_var_keyword=False,
+            ),
+            id='instance.g'
+        ),
+        pytest.param(
+            d.h,
+            ArgumentSpecification(
+                arguments=tuple([]),
+                has_var_positional=False,
+                has_var_keyword=False,
+            ),
+            id='instance.h'
+        ),
     ]
 )
-def test_arg_spec(func, expected):
-    assert expected == get_arguments_specification(func)
+def test_arg_spec(func, expected: ArgumentSpecification):
+    result = get_arguments_specification(func)
+    assert isinstance(result, ArgumentSpecification)
+    assert expected.has_var_positional == result.has_var_positional
+    assert expected.has_var_keyword == result.has_var_keyword
+
+    for expected_arg, result_arg in zip(expected.arguments, result.arguments):
+        assert expected_arg.name == result_arg.name
+        assert expected_arg.has_default == result_arg.has_default