+++ CHANGELOG.rst
@@ -1,7 +1,18 @@
 Changelog
 =========
 
+
+0.3.1-dev
+---------
+
+Bug fixes
+^^^^^^^^^
+
+- When :code:`DependencyManager.factory` was applied on a class it would return
+  an instance of it and not the class itself.
+
+
 0.3.0 (2018-04-29)
--------------------
+------------------
 
 Initial release